+++ src/tracks/AbstractIntervalFeature.java
@@ -1,0 +1,15 @@
+package tracks;
+
+abstract class AbstractIntervalFeature {
+
+	private String bgColor= null;
+
+	protected String getBgColor() {
+		return bgColor;
+	}
+
+	protected void setBgColor(String bgColor) {
+		this.bgColor = bgColor;
+	}
+	
+}