+++ CHANGELOG.md
@@ -1,7 +1,20 @@
+New in 1.9.0
+============
+
+* New command `featureColorForRegex` sets custom colour of individual features.
+
+* Add command `PERCENT [PERCENT]` to zoom into a region on the current 
+window. E.g. `.25 .5` moves to the second quarter of the current 
+window. This command deprecates the use of suffix `c` in command command `INT [INT]`. 
+
+* On exit via `q` command the screen is cleared. This avoids leaving the the screen in mixed colours. 
+
+* Fixed some minor bugs.
+
 New in 1.8.0
 ============
 
-* Transcript names (or in general feature names) for GTF files ca be set via `gffNameAttr` (Fix issue 74).
+* Transcript names (or in general feature names) for GTF files can be set via `gffNameAttr` (Fix issue 74).
 
 * There should be a mild improve in speed, especially at start up, due to building the
   jar with libraries *extracted* instead of *packaged*.